refactor(shaderpark): extract completion module
